# config/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Watch & Do API",
        "description": "Watch & Do API",
        "contact": {
            "name": "Alexandre Maia",
            "url": "http://wwwalexandremaia.pt",
            "email": "info@alexandremaia.pt"
        },
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/users/{id}": {
            "put": {
                "tags": [ "Users" ],
                "description": "<h2>Update User</h2><br><p>Register Users can update her own info</p><p>Admin Users can update Register Users info and her own.</p><p>Super Admin can update all Users</p><br><p>Role required: <strong>Register</strong>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#components/schemas/Users" },
                            "examples": {
                                "User": {
                                    "value": {
                                        "username": "userupdate",
                                        "email": "userupdate@mail.pt"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "examples": {
                                "Success": {
                                    "value": "User Updated"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
...
        "/wd/users": {
            "get": {
                "tags": [ "Users" ],
                "description": "<h2>Get Users List<h2><br><p>Role required: <strong>Admin</strong>",
                "responses": {
                    "200": {
                        "description": "Users List",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Users" },
                                "examples": {
                                    "Users List": [
                                        {
                                            "id": "1",
                                            "username": "Daniel",
                                            "email": "daniel@mail.pt",
                                            "status": "1",
                                            "role_id": "1"
                                        },
                                        {
                                            "id": "2",
                                            "username": "Joaquim",
                                            "email": "joaquim@mail.pt",
                                            "status": "1",
                                            "role_id": "1"
                                        }
                                    ]
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/users/{id}": {
            "get": {
                "tags": [ "Users" ],
                "description": "Get User Info by Id<br>Role required Admin",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Info",
                        "content": {
                            "application/json": {
                                "schema": { "$ref": "#components/schemas/Users" },
                                "examples": {
                                    "User": {
                                        "id": "1",
                                        "username": "Daniel",
                                        "email": "daniel@mail.pt",
                                        "status": "1",
                                        "role_id": "1"
                                    }
                                }
                            },
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            },
            "put": {
                "tags": [ "Users" ],
                "description": "<h2>Update User</h2><br><p>Only Super Admin can update data from other Super Admin</p><p>Role required: <strong>Admin</strong>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#components/schemas/Users" },
                            "examples": {
                                "User": {
                                    "value": {
                                        "username": "userupdate",
                                        "email": "userupdate@mail.pt"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "examples": {
                                "Success": {
                                    "value": "User Updated"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            },
            "delete": {
                "tags": [ "Users" ],
                "description": "<h2>Delete User</h2><br><p>User can't delete himself.</p><p>Role required: <strong>Super Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "examples": {
                                "Success": {
                                    "value": "User Deleted"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Can't delete yourself"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/users/{id}/role": {
            "put": {
                "tags": [ "Users" ],
                "description": "<h2>Change User Role</h2><br><p>Super Admin can only assign Roles to others Users excluding him self.</p><p>Role required: <strong>Super Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#components/schemas/Users" },
                            "examples": {
                                "Register": {
                                    "value": {
                                        "role_id": "1"
                                    }
                                },
                                "Admin": {
                                    "value": {
                                        "role_id": "2"
                                    }
                                },
                                "Super Admin": {
                                    "value": {
                                        "role_id": "3"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "examples": {
                                "Success": {
                                    "value": "Password Changed"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Can't delete yourself"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
        "/wd/users/{id}/status": {
            "put": {
                "tags": [ "Users" ],
                "description": "<h2>Change User Status</h2><br><p>Can't change own status</p><p>Only Super Admin can change other Super Admin Status<p></p><p>Role required: <strong>Admin</strong></p>",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "examples": {
                                "Success": {
                                    "value": "User Status Active<br>User Status Inactive"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Error": {
                                        "value": "Error"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Invalid Token"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                },
                                "examples": {
                                    "Unauthorized": {
                                        "value": "Can't change your own Status"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [ { "bearer": [] } ]
            }
        },
...
    }
}